% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html.R
\name{add_html}
\alias{add_html}
\title{Add HTML elements}
\usage{
add_html()
}
\value{
Write lines of html codes for the desired objects where the function was called.
}
\description{
Gadget to insert various objects in HTML documents
}
\author{
Nicolas Mangin
}
