% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_yaml.R
\name{add_yaml}
\alias{add_yaml}
\title{Gadget to insert the yaml of a research paper.}
\usage{
add_yaml()
}
\value{
Write lines of the YAML at the top of the Rmd files in which the paper is written.
}
\description{
Gadget to insert the yaml of a research paper.
}
