% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.R
\name{aov_r}
\alias{aov_r}
\title{ANOVA for factorial designs}
\usage{
aov_r(
  data,
  response,
  between = NULL,
  within = NULL,
  id,
  type = 3,
  es = "eta",
  sphericity = "auto",
  markdown = TRUE
)
}
\arguments{
\item{data}{Your dataset in long format, can have some missing values.}

\item{response}{Response variable, numeric.}

\item{between}{Quoted or unquoted variable indicating the between-subject(s) factor(s)/column(s) in data. Default is NULL indicating no between-subjects factors. Must be character vector if more than one between-subject(s) factor(s)/column(s) is specified.}

\item{within}{Quoted or unquoted variable indicating the within-subject(s) factor(s)/column(s) in data. Default is NULL indicating no between-subjects factors. Must be character vector if more than one within-subject(s) factor(s)/column(s) is specified.}

\item{id}{Quoted or unquoted variable (of length 1) indicating the subject identifier column in data.}

\item{type}{The type of sums of squares for the ANOVA. Possible values are "II", "III", 2, or 3 (default).}

\item{es}{The effect size used to estimate the effects of the factors on the response variable. Possible values are 'omega' or 'eta' (default).}

\item{sphericity}{If \code{"none"}, then sphericity assumption is assumed to be met for within-subject(s) factor(s). "GG": applies Greenhouse-Geisser correction. "HF": applies Hyunh-Feldt correction. 'auto' (Default) choose the appropiate correction based on Mauchly test of sphericity (p-value > 0.05)}

\item{markdown}{Whether you want the output formated for inline R Markdown or as plain text.}
}
\value{
A list with full statistical results, estimates and simple stats in APA style for each factor and interaction term(s).
}
\description{
This is function let you perform automated inferential testing based on certain assumptions, some of which are tested automatically, then the propper test is perform, giving you an APA formated output with your statistical results.
}
\keyword{aov_r}
