% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{as.data.frame.writR}
\alias{as.data.frame.writR}
\title{as.data.frame method for writR objects}
\usage{
\method{as.data.frame}{writR}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A writR object from any of one_sample, two_sample, k_sample, autest or contingency.}

\item{row.names}{Exported from other methods.}

\item{optional}{Exported from other methods.}

\item{...}{Currently ignored}
}
\description{
as.data.frame method for writR objects
}
