% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\docType{data}
\name{charts}
\alias{charts}
\title{List of charts which can be inserted}
\format{A data frame with the following variables:
\itemize{
\item information
\item object
\item output
\item type
\item render
\item use
}}
\usage{
data("charts")
}
\description{
A list of charts which can be inserted through the function chart_insert
}
\keyword{charts}
