% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Data table without NA's and with rowid (paired designs only).}
\usage{
clean_data(data, x, y, rowid = NULL, paired = FALSE, wide = FALSE, ...)
}
\arguments{
\item{data}{Data frame from which \code{x} and \code{y} (and possibly \code{rowid} if provided) will be searched.}

\item{x}{Unquoted name for the grouping factor. Must be present in data}

\item{y}{Unquoted name for the response variable. Must be present in data.}

\item{rowid}{Unquoted name for the subject-id column. If null, then is assumed that data is sorted for paired designs, creating one. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate when there are more than two levels in x and there are NAs present.}

\item{paired}{Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is \code{FALSE.}}

\item{wide}{Logical to whether return a data.frame in wide format (\code{TRUE}, i.e. one columns per group/time) or in long format (\code{FALSE}).}

\item{...}{Currently ignored.}
}
\description{
Data table without NA's and with rowid (paired designs only).
}
