% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency.R
\name{contingency}
\alias{contingency}
\title{Report categorical analyses}
\usage{
contingency(
  data,
  x,
  y = NULL,
  paired = FALSE,
  ratio = NULL,
  conf.level = 0.95,
  character.only = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame containing the variables \code{x} and \code{y}.}

\item{x}{Factor variable, quoted or unquoted.}

\item{y}{Factor. If \code{NULL}, a goodness-of-fit is carried, otherwise a two-way analysis is performed.}

\item{paired}{Logical. If \code{TRUE} McNemar's Chi-squared test is carried on.}

\item{ratio}{A vector of probabilities of the same length of x. An error is given if any entry of p is negative.}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to 0.95 (95\%).}

\item{character.only}{Logical. checks whether to use the unevaluated expression or its
content (when TRUE), asumming is a character vector. Defaults to \code{FALSE}.}

\item{...}{Currently not used.}
}
\description{
Perform nominal/ordinal analysis on 1 dimensional table for goodnes-of-fit chi-squared, and two dimensional data for Pearson chi-squared, Fisher exact test or (if paired), McNemar test reporting their corresponding stats and effect sizes in APA Style.
}
\keyword{contingency}
