% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_words.R
\name{count_words}
\alias{count_words}
\title{Count the number of words in a text.}
\usage{
count_words(text = "")
}
\arguments{
\item{text}{Character. The sentence(s) in which the words have to be counted.}
}
\value{
Numeric. Number of words in the text.
}
\description{
Count the number of words in a text.
}
