% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_jelcodes.R
\docType{data}
\name{dat_jelcodes}
\alias{dat_jelcodes}
\title{List of JEL codes}
\format{
A data frame with the following variables:
\itemize{
\item level1
\item level2
\item level3
\item code
}
}
\source{
https://www.aeaweb.org/econlit/jelCodes.php?view=jel
}
\usage{
data("dat_jelcodes")
}
\description{
A list of codes from the jounal of economic literature
}
\author{
Nicolas Mangin
}
\keyword{jelcodes}
