% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl.R
\name{dbl}
\alias{dbl}
\title{dbl format as a number with two digits.}
\usage{
dbl(x)
}
\arguments{
\item{x}{Double or Integer. Number to be properly formated.}
}
\value{
Character. Formated number.
}
\description{
dbl format as a number with two digits.
}
