% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparser.R
\name{deparser}
\alias{deparser}
\title{Get the deparsed value of the outermost expression}
\usage{
deparser(x, character.only = FALSE, ...)
}
\arguments{
\item{x}{An expression from which to capture the outermost expression.}

\item{character.only}{whether to treat \code{x} as a character. Default is FALSE.}

\item{...}{Currently not used.}
}
\description{
Get the deparsed value of the outermost expression
}
