% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_appendix.R
\name{gen_appendix}
\alias{gen_appendix}
\title{Format appendices.}
\usage{
gen_appendix(..., keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"))
}
\arguments{
\item{...}{Arguments passed to \code{pdf_document()}.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown:render]{render}}
}
\description{
Actually make sure no special format is applied to the appendices so that the main template applies.
}
