% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_paper.R
\name{gen_paper}
\alias{gen_paper}
\title{Format submissions}
\usage{
gen_paper(
  ...,
  engine = "xelatex",
  journal = "aom",
  keep_tex = TRUE,
  citation_package = "natbib",
  md_extensions = c("-autolink_bare_uris")
)
}
\arguments{
\item{...}{Arguments passed to \code{pdf_document()}.}

\item{engine}{character. Latex engine to be used to produce the pdf.}

\item{journal}{character. Journal to which the paper is submitted. See details for the available options.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown:render]{render}}
}
\description{
Format submissions to various academic conferences and journals.
}
\details{
The templates currently available are:
\itemize{
\item aos for Accounting, Organizations, and Society
\item bria for Behavioral Research in Accounting
\item car for Contemporary Accounting Research
\item cpa for Critical Perspectives on Accounting
\item ear for European Accounting Review
\item jae for Journal of Accounting and Economics
\item jar for Journal of Accounting Research
\item jaaf for Journal of Accounting, Auditing, and Finance
\item jmar for Journal of Management Accounting Research
\item mar for Management Accounting Research
\item tar for The Accounting Review
\item jfe for Journal of Financial Economics
\item amj for Academy of Management Journal
\item amr for Academy of Management Review
\item jibs for Journal of International Business Studies
\item jom for Journal of Management
\item mnsc for Management Science
\item orsc for Organization Science
\item smj for Strategic Management Journal
\item orm for Organizational Research Methods
\item jap for Journal of Applied Psychology
}
}
\author{
Nicolas Mangin
}
