% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{GG}
\alias{GG}
\title{Greenhouse-Geisser epsilon}
\usage{
GG(model)
}
\arguments{
\item{model}{A repeated measures ANOVA model using afex.}
}
\description{
Internal function inside \code{sphericity_check}. Returns Greenhouse-Geisser epsilon for corrections for departure from sphericity
}
