% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_place_axis.R
\name{graph_place_axis}
\alias{graph_place_axis}
\title{Add axis at the specified coordinate on a ggplot2 graph}
\usage{
graph_place_axis(graph, x = 0, y = 0, xunit = 0.5, yunit = 0.5)
}
\arguments{
\item{graph}{ggplot2 graph.}

\item{x}{Numeric. Where the x axis should appear.}

\item{y}{Numeric. Where the y axis should appear.}

\item{xunit}{Numeric. Units of the x axis (for breaks)}

\item{yunit}{Numeric. Units of the y axis (for breaks)}
}
\value{
A formated graph.
}
\description{
Add axis at the specified coordinate on a ggplot2 graph
}
\examples{
library(ggplot2)
library(dplyr)
library(writR)
g <- iris \%>\%
  mutate_if(is.numeric, scale) \%>\%
  ggplot(aes(x = Sepal.Length, y = Petal.Length, color = Species)) +
  geom_point() +
  geom_smooth(method = "lm") +
  scale_colour_grey() +
  scale_x_continuous(breaks = seq(-2, 10, 0.5)) + xlab("Setal") +
  scale_y_continuous(breaks = seq(-2, 10, 0.5)) + ylab("Pepal") +
  graph_theme(axis = FALSE, legend.position = "bottom")
  graph_place_axis(g, x = 0, y = 0, xunit = 0.5, yunit = 0.5)
}
