% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theme.R
\name{graph_theme}
\alias{graph_theme}
\title{Apply a common theme to graphs generated with ggplot2}
\usage{
graph_theme(base_size = 11, base_family = "",
  legend.position = "bottom", legend.box = NULL, axis = TRUE,
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{Numeric. Adjust size parameters.}

\item{base_family}{Character. Font.}

\item{legend.position}{Character. "none", "bottom", "top", or vector of coordinates.}

\item{legend.box}{Character. "horizontal" or "vertical".}

\item{axis}{Logical. Whether axis should be dranw on the side.}

\item{base_line_size}{Numeric. Adjust size parameters.}

\item{base_rect_size}{Numeric. Adjust size parameters.}
}
\value{
A formated graph.
}
\description{
Apply a common theme to graphs generated with ggplot2
}
\examples{
library(ggplot2)
library(dplyr)
library(writR)
g <- iris \%>\%
  mutate_if(is.numeric, scale) \%>\%
  ggplot(aes(x = Sepal.Length, y = Petal.Length, color = Species)) +
  geom_point() +
  geom_smooth(method = "lm") +
  scale_colour_grey() +
  scale_x_continuous(breaks = seq(-2, 10, 0.5)) + xlab("Setal") +
  scale_y_continuous(breaks = seq(-2, 10, 0.5)) + ylab("Pepal") +
  graph_theme(axis = FALSE, legend.position = "bottom")
  graph_place_axis(g, x = 0, y = 0, xunit = 0.5, yunit = 0.5)
}
