% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HF}
\alias{HF}
\title{Hyunh-Feldt epsilon}
\usage{
HF(model, gg = NULL)
}
\arguments{
\item{model}{A repeated measures ANOVA model using afex.}

\item{gg}{Greenhouse-Geisser epsilon. Used for posterior calculations. If is not supplied it will be calculated.}
}
\description{
Internal function inside \code{sphericity_check}. Returns Hyunh-Feldt epsilon for corrections for departure from sphericity
}
