% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_box.R
\name{html_box}
\alias{html_box}
\title{Create a HTML box which can be placed anywhere on the page or slide, just by specifying the distance from the top left.}
\usage{
html_box(
  text = "",
  width = "20\%",
  top = "0px",
  left = "0px",
  background = rgb(0, 0, 0, 0.75),
  color = "#FFF",
  size = 14,
  align = "justify"
)
}
\arguments{
\item{text}{Character. Content of the box.}

\item{width}{Character. Width of the box.}

\item{top}{Character. Distance from the top.}

\item{left}{Character. Distance from the right.}

\item{background}{Character. Background color.}

\item{color}{Character. Font color.}

\item{size}{Integer. Font size.}

\item{align}{Character. Text alignment}
}
\value{
Character. Formatted and placed box.
}
\description{
Create a HTML box which can be placed anywhere on the page or slide, just by specifying the distance from the top left.
}
