% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_layout.R
\name{html_layout}
\alias{html_layout}
\title{Create an html grid layout}
\usage{
html_layout(layout = list(c(1, 2, 1), c(3, 1), c(1, 1, 1, 1)))
}
\arguments{
\item{layout}{list of vectors. one vector per row, the integers in the vector specifying the relative width of rach cell of this row.}
}
\value{
Character. html grid.
}
\description{
Create an html grid layout
}
