% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_row.R
\name{html_row}
\alias{html_row}
\title{Gather html cells on a row; it is meant to be used through html_layout.}
\usage{
html_row(cells)
}
\arguments{
\item{cells}{character vector. Vector of outputs of the function html_cell}
}
\value{
Character. html row composed of several cells.
}
\description{
Gather html cells on a row; it is meant to be used through html_layout.
}
\seealso{
html_layout
}
