% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_slide.R
\name{html_slide}
\alias{html_slide}
\title{Create a HTML slide}
\usage{
html_slide(
  title = "",
  colorback = "#FFFFFF",
  transback = "slide",
  transdata = "fade",
  image = "",
  video = ""
)
}
\arguments{
\item{title}{Character. Title of the slide.}

\item{colorback}{Character. Background color.}

\item{transback}{Character. Type of transition for the background: fade, slide, convex, concave, or zoom.}

\item{transdata}{Character. Type of transition for data: fade, slide, convex, concave, or zoom.}

\item{image}{Character. Address of the background image.}

\item{video}{Character. Address of the background video.}
}
\value{
Character. Reveal js slide formatted for rmarkdown.
}
\description{
Insert a reveal js slide
}
\author{
Nicolas Mangin
}
