% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_star_rating.R
\name{html_star_rating}
\alias{html_star_rating}
\title{Function inserting a star rating in an html document. This requires the css and the starrating.js to be specified separately in the header and footer of the html document.}
\usage{
html_star_rating(source = "", page = 1)
}
\arguments{
\item{source}{Character. Name of the book.}

\item{page}{Integer. Number of the section in the book.}
}
\description{
Function inserting a star rating in an html document. This requires the css and the starrating.js to be specified separately in the header and footer of the html document.
}
