% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{int}
\alias{int}
\title{int formats as an integer.}
\usage{
int(x)
}
\arguments{
\item{x}{Double or Integer. Number to be properly formated.}
}
\value{
Character. Formated number.
}
\description{
int formats as an integer.
}
