% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{is_var.equal}
\alias{is_var.equal}
\title{Levene's Test}
\usage{
is_var.equal(y, x, alpha = 0.05, center = stats::median)
}
\arguments{
\item{y}{Numeric vector (response).}

\item{x}{Grouping factor.}

\item{alpha}{Threshold for null hipotesis (of normality) rejection.}

\item{center}{A function to compute the center of each group; mean gives the original Levene's test; the default, median, provides a more robust test.}
}
\description{
Levene's Test
}
