% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jelcodes.R
\docType{data}
\name{jelcodes}
\alias{jelcodes}
\title{List of JEL codes}
\format{A data frame with the following variables:
\itemize{
\item level1
\item level2
\item level3
\item code
}}
\source{
https://www.aeaweb.org/econlit/jelCodes.php?view=jel
}
\usage{
data("jelcodes")
}
\description{
A list of codes from the jounal of economic literature
}
\keyword{jelcodes}
