% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journals.R
\docType{data}
\name{journals}
\alias{journals}
\title{List of journals with available templates}
\format{
A data frame with the following variables:
\itemize{
\item field Character. The main discipline of the journal.
\item acronym Character. Acronym of the journal.
\item journal Character. Full name of the journal.
\item template Character. Name of the Latex template to be used in the paper generation.
}
}
\usage{
data("journals")
}
\description{
A list of journals with the corresponding templates
}
\keyword{journals}
