% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_sample.R
\name{k_sample}
\alias{k_sample}
\title{K Sample test}
\usage{
k_sample(
  data,
  x,
  y,
  rowid = NULL,
  type,
  paired = FALSE,
  var.equal = FALSE,
  is_spherical = NULL,
  adjust = NULL,
  effsize.type = "unbiased",
  conf.level = 0.95,
  character.only = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame from which \code{x} and \code{y} (and possibly \code{rowid} if provided) will be searched.}

\item{x}{Character for the grouping factor. Must be present in data}

\item{y}{Character for the response variable. Must be present in data.}

\item{rowid}{Character for the subject-id column. If null, then is assumed that data is sorted for paired designs, creating one. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate when there are more than two levels in x and there are NAs present.}

\item{type}{Missing (default) or \code{NULL} for checking the normality and homogeneity of variances for test selection. Other options are \code{"p"} for parametric, \code{"np"} for non-parametric and \code{"r"} for robust tests.}

\item{paired}{Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is \code{FALSE.}}

\item{var.equal}{Logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{is_spherical}{Logical. checks whether to assume that the sphericity assumptions holds or not, if \code{NULL} (the default) it will be tested using mauchly's test with a threshold of 0.05.}

\item{adjust}{Character. correction for sphericity to be applied, it can be any character of length one starting with 'g' (indicating Greenhouse–Geisser correction) or 'h' (indicating Huynh–Feldt correction).}

\item{effsize.type}{Options are \code{"unbiased"} or \code{"omega"} for partial omega squared and \code{"biased"} or \code{"eta"} for partial eta squared as a measure of effect size. For non-parametric analysis, Kendalls' W is used for paired designs, where rank epsilon squared is used for independent groups designs.}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to 0.95 (95\%).}

\item{character.only}{Logical. checks whether to use the unevaluated expression or its
content (when TRUE), asumming is a character vector. Defaults to \code{FALSE}.}

\item{...}{Currently ignored.}
}
\description{
A list containing results from a multi-sample test.
}
