% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lablr.R
\name{lablr}
\alias{lablr}
\title{Generate report-ready labels for statistical report}
\usage{
lablr(t, markdown = FALSE)
}
\arguments{
\item{t}{Output from any of the functions, k_sample, two_sample or one_sample.}

\item{markdown}{Logical (default FALSE), indicating if the report-ready labels should be formated for inline code for R markdown (using mathjax and markdown syntax), or if the output should be in plain text (the default).}
}
\description{
A list containing stats, p value, effectsize, confidence/credible interval and a concatenated string named 'full'.
}
