% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_env.R
\name{latex_env}
\alias{latex_env}
\title{Insert text (question, proposition, hypothesis, quote) in Latex environments}
\usage{
latex_env(
  text = "",
  env = "hypothesis",
  space_before = "15pt",
  space_between = "5pt",
  space_after = "15pt",
  author = "",
  bibkey = FALSE,
  pages = ""
)
}
\arguments{
\item{text}{charactervector. Text to be inserted. If more than one entry are prvided, create a sub-environment.}

\item{env}{Character. Latex environment in which the text has to be inserted.}

\item{space_before}{Character. Spacing before the group.}

\item{space_between}{Character. Spacing between entries.}

\item{space_after}{Character. Spacing after the group.}

\item{author}{Character. Person quoted.}

\item{bibkey}{Logical. Wheter the author is formated as a reference.}

\item{pages}{Character. Pages of the quote.}
}
\value{
\preformatted{         Character. Formated Latex environment.
}
}
\description{
Insert text (question, proposition, hypothesis, quote) in Latex environments
}
