% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_fit_stargazer.R
\name{latex_fit_stargazer}
\alias{latex_fit_stargazer}
\title{Create a Latex tabular environment for stargazer and insert it in an adjustable table}
\usage{
latex_fit_stargazer(
  ...,
  table_caption = "",
  table_label = "",
  env = "table",
  adjwidth = 0.75,
  adjheight = 0.75,
  space_before = "15pt",
  space_after = "15pt"
)
}
\arguments{
\item{...}{Arguments passed to the stargazer function.}

\item{table_caption}{Character. Title of the table.}

\item{table_label}{Character. Label of the table (for referencing in Latex).}

\item{env}{Character. One of "table" or "sidewaystable".}

\item{adjwidth}{Double. Percentage of the width of the text area.}

\item{adjheight}{Double. Percentage of the height of the text area.}

\item{space_before}{Character. Spacing before the table.}

\item{space_after}{Character. Spacing after the table.}
}
\value{
\preformatted{         A printable Latex table
}
}
\description{
Create a Latex tabular environment for stargazer and insert it in an adjustable table
}
