% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_long_stargazer.R
\name{latex_long_stargazer}
\alias{latex_long_stargazer}
\title{Create a longtable stargazer output}
\usage{
latex_long_stargazer(
  ...,
  table_caption,
  table_label,
  threeparttable = TRUE,
  landscape = FALSE,
  font_size = "small"
)
}
\arguments{
\item{...}{Arguments passed to the stargazer function.}

\item{table_caption}{Character. Title of the table.}

\item{table_label}{Character. Label of the table (for referencing in Latex).}

\item{threeparttable}{Logical. Align captions with table.}

\item{landscape}{Logical. Put on a landscape page.}

\item{font_size}{Integer. Size of the font inside the table.}
}
\value{
\preformatted{           A printable Latex long table
}
}
\description{
Create a longtable stargazer output
}
