% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_panels.R
\name{latex_panels}
\alias{latex_panels}
\title{Create Latex Panels}
\usage{
latex_panels(
  panels,
  subcaptions,
  tab_width = c(1),
  end_line = c(F),
  maincaption,
  label,
  orientation = "portrait",
  font_size = "normalsize"
)
}
\arguments{
\item{panels}{Character vector. Output of the capture.output function applied to latex tables.}

\item{subcaptions}{character. Titles of the subtables.}

\item{tab_width}{Character. Width of the minipage if length == 1, width of each subtable otheriwse.}

\item{end_line}{Numeric. Vector of same length as tab_width; 1 if the subtable should end the line, 0 otherwise.}

\item{maincaption}{Character. Title of the group of figures.}

\item{label}{Character. Label of the group of figures.}

\item{orientation}{Character. "portrait" of "landscape".}

\item{font_size}{Character. Latex specification of the font size (e.g. "normalsize", "small", "tiny").}
}
\value{
Character. A set of objects organized as panels.
}
\description{
Insert tables or figures in panels in a Latex environments
}
\author{
Nicolas Mangin
}
