% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{mauchly}
\alias{mauchly}
\title{Mauchly's Test of Sphericity}
\usage{
mauchly(model)
}
\arguments{
\item{model}{A repeated measures ANOVA model using afex.}
}
\description{
Internal function inside \code{sphericity_check}. Return a p-value from Mauchly's Test for within-subjects factors.
}
