% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample.R
\name{one_sample}
\alias{one_sample}
\title{One Sample test}
\usage{
one_sample(
  data,
  y,
  type = "auto",
  test.value = 0,
  trim = 0.2,
  nboot = 100L,
  effsize.type = "unbiased",
  alternative = "two.sided",
  conf.level = 0.95,
  lbl = if (is.null(markdown)) FALSE else TRUE,
  markdown = NULL,
  ...
)
}
\arguments{
\item{data}{data frame from which \code{x} and \code{y} (and possibly \code{rowid} if provided) will be searched.}

\item{y}{Character for the response variable Must be present in data}

\item{type}{set \code{"auto"} (default) for checking the normality for test selection. Other options are \code{"p"} for parametric, \code{"np"} for non-parametric and \code{"r"} for robust tests.}

\item{test.value}{A number indicating the true value of the mean (Default: 0) to be tested.}

\item{trim}{Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.}

\item{nboot}{Number of bootstrap samples for computing confidence interval for the effect size (Default: 100L).}

\item{effsize.type}{options are \code{"unbiased"} or \code{"g"} for Hedges g and \code{"biased"} or \code{"d"} for Cohen's d as a measure of effect size. rank-biserial correlation is used for non-parametric analysis.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to 0.95 (95\%).}

\item{lbl}{Logical (default FALSE) indicating if a report ready output is desired. This will change the output to a list with characters rather than numeric vectors.}

\item{markdown}{Logical (default FALSE). If \code{lbl} is TRUE, then this argument specify if the report-ready labels should be formated for inline code for R markdown (using mathjax and markdown syntax), or if the output should be in plain text (the default).}

\item{...}{Currently ignored.}
}
\description{
A list containing results from a one-sample test.
}
