% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.R
\name{pct}
\alias{pct}
\title{pct transforms numbers between 0 and 1 in percentages.}
\usage{
pct(x)
}
\arguments{
\item{x}{Double. Number to be properly formated.}
}
\value{
Character. Formated number.
}
\description{
pct transforms numbers between 0 and 1 in percentages.
}
