% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{print.writR}
\alias{print.writR}
\title{Print method for writR objects}
\usage{
\method{print}{writR}(x, ...)
}
\arguments{
\item{x}{A writR object from any of one_sample, two_sample, k_sample, autest or contingency.}

\item{...}{Currently ignored}
}
\description{
Print method for writR objects
}
