% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_anova.R
\name{rm_anova}
\alias{rm_anova}
\title{Repeated measures ANOVA}
\usage{
rm_anova(
  data,
  x,
  y,
  is_spherical,
  adjust,
  effsize.type = "omega",
  conf.level = 0.95,
  character.only = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame from which \code{x} and \code{y} (and possibly \code{rowid} if provided) will be searched.}

\item{x}{Character name for the grouping factor. Must be present in data}

\item{y}{Character name for the response variable. Must be present in data.}

\item{is_spherical}{Logical. checks whether to assume that the sphericity assumptions holds or not, if missing (the default) it will be tested using mauchly test with a threshold of 0.05.}

\item{adjust}{Character. correction for sphericity to be applied, it can be any character of length one starting with 'g' (indicating Greenhouse–Geisser correction) or 'h' (indicating Huynh–Feldt correction).}

\item{effsize.type}{Options are \code{"unbiased"} or \code{"omega"} for partial omega squared and \code{"biased"} or \code{"eta"} for partial eta squared as a measure of effect size. For non-parametric analysis, Kendalls' W is used for paired designs, where rank epsilon squared is used for independent groups designs.}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to 0.95 (95\%).}

\item{character.only}{Logical. checks whether to use the unevaluated expression or its
content (when TRUE), asumming is a character vector. Defaults to \code{FALSE}.}

\item{...}{Currently not used.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This functions allows you to perform a one way repeated measures ANOVA,
making adjustments for violations of sphericity, with no dependencies
related to \code{{afex}} or \code{{car}} packages.
}
