% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{sci}
\alias{sci}
\title{sci format scientifically.}
\usage{
sci(x)
}
\arguments{
\item{x}{Double or Integer. Number to be properly formated.}
}
\value{
Character. Formated number.
}
\description{
sci format scientifically.
}
