% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_glossary.R
\name{sort_glossary}
\alias{sort_glossary}
\title{Sort Book Glossary}
\usage{
sort_glossary(glossary)
}
\arguments{
\item{glossary}{character string. Address of the glossary to sort.}
}
\description{
Sort the entries of a glossary at the end of a PDF book.
}
\author{
Nicolas Mangin
}
