% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sphericity}
\alias{sphericity}
\title{Mauchly's Test of Sphericity}
\usage{
sphericity(
  object,
  Sigma = diag(nrow = p),
  T = Thin.row(Proj(M) - Proj(X)),
  M = diag(nrow = p),
  X = ~0,
  idata = data.frame(index = seq_len(p))
)
}
\arguments{
\item{object}{object of class SSD or mlm}

\item{Sigma}{matrix to be proportional to.}

\item{T}{transformation matrix. By default computed from M and X.}

\item{M}{formula or matrix describing the outer projection.}

\item{X}{formula or matrix describing the inner projection.}

\item{idata}{data frame describing intra-block design.}
}
\description{
Low-level function behind \code{mauchly.test()}
}
