% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sphericity_check}
\alias{sphericity_check}
\title{Suggested sphericity correction for repeated measures ANOVA}
\usage{
sphericity_check(model)
}
\arguments{
\item{model}{A repeated measures ANOVA model using Afex.}
}
\description{
Internal function inside \code{k_sample}. Return the Spherecity correction suggested based on Mauchly test in one-way repeated measures designs
}
