% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_bivariate.R
\name{stat_bivariate}
\alias{stat_bivariate}
\title{Generate and format a correlation table for publication.}
\usage{
stat_bivariate(
  x,
  use = "pairwise",
  method = "pearson",
  digits = 3,
  variables = NULL,
  addpval = T
)
}
\arguments{
\item{x}{Tibble or dataframe. Numeric variables as columns, observations as rows.}

\item{use}{Character. "pairwise" is the default value and will do pairwise deletion of cases. "complete" will select just complete cases.}

\item{method}{Character. "pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall"}

\item{digits}{Numeric. Number of digits to be displayed in the formated table.}

\item{variables}{Character vector. Names of the variables in rows (if different from names in the dataframe).}

\item{addpval}{Logical. Whether p-values should be added below the correlations.}
}
\value{
\preformatted{        A correlation table formated for publication.
}
}
\description{
Generate and format a correlation table for publication.
}
