% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_stars.R
\name{stat_stars}
\alias{stat_stars}
\title{Add stars to a vector of coefficients based on the p-value.}
\usage{
stat_stars(
  coeff = NULL,
  pval = NULL,
  thresholds = c(0.01, 0.05, 0.1),
  digits = 3
)
}
\arguments{
\item{coeff}{Numeric vector. Coefficients.}

\item{pval}{Numeric vector. Corresponding p-values.}

\item{thresholds}{Numeric vector. three values indicating the cut-off p-values.}

\item{digits}{Integer. Number of digits for rounding.}
}
\value{
Character vector with coefficients formatted for publication.
}
\description{
Add stars to a vector of coefficients based on the p-value.
}
