% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_totals.R
\name{stat_totals}
\alias{stat_totals}
\title{Add totals (sums or means) to a dataframe.}
\usage{
stat_totals(x, rows = TRUE, columns = TRUE, omit_col = NA, summary = "sum")
}
\arguments{
\item{x}{Dataframe. Data to which you wish to add the totals}

\item{rows}{Logical. Whether totals should be added at the bottom of the table.}

\item{columns}{Logical. Whether totals should be added at the right-side of the table.}

\item{omit_col}{Character vector. Names of the variables not to be included in the total (for instance containing the id of the observation).}

\item{summary}{Character. Whether the reported statistic should be a "sum" or a "mean".}
}
\value{
Tibble with the summary statistics added as a column or a row.
}
\description{
Add totals (sums or means) to a dataframe.
}
