% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_univariate.R
\name{stat_univariate}
\alias{stat_univariate}
\title{Generate a table displaying univariate statistics for all the variables included in x which must be numeric.}
\usage{
stat_univariate(
  x,
  variables = NULL,
  include = c("Variable", "Count", "Min", "Median", "Mean", "Max", "Range", "St.Dev",
    "Skew", "Kurt")
)
}
\arguments{
\item{x}{Tibble or dataframe. Table with only numeric variables.}

\item{variables}{Character vector. Names of the variables.}

\item{include}{Character vector. Names of the statistics to report: "Count", "Min", "Median", "Mean", "Max", "Range", "St.Dev","Skew" and "Kurt".}
}
\value{
Dataframe. Table of univariate statistics.
}
\description{
Generate a table displaying univariate statistics for all the variables included in x which must be numeric.
}
