% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{style.p}
\alias{style.p}
\title{Style p value}
\usage{
style.p(p, k = 3)
}
\arguments{
\item{p}{Vector of p value(s)}

\item{k}{Number of decimal places}
}
\value{
Character vector of the same length of the input.
}
\description{
Style p value for pretty printing
}
