% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparser.R
\name{subs}
\alias{subs}
\title{Always returns the outermost expression}
\source{
\url{https://stackoverflow.com/a/26558733}
}
\usage{
subs(x, env, ...)
}
\arguments{
\item{x}{An expression from which to capture the outermost expression.}

\item{env}{The last environment in which to search \code{x}. Default is missing.}

\item{...}{Currently not used.}
}
\description{
Always returns the outermost expression
}
