% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_bivariate.R
\name{table_bivariate}
\alias{table_bivariate}
\title{Report bivariate statistics}
\usage{
table_bivariate(
  x,
  use = "pairwise",
  method = "pearson",
  digits = 3,
  variables = NULL,
  addpval = T
)
}
\arguments{
\item{x}{Tibble or dataframe. Numeric variables as columns, observations as rows.}

\item{use}{Character. "pairwise" is the default value and will do pairwise deletion of cases. "complete" will select just complete cases.}

\item{method}{Character. "pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall"}

\item{digits}{Numeric. Number of digits to be displayed in the formated table.}

\item{variables}{Character vector. Names of the variables in rows (if different from names in the dataframe).}

\item{addpval}{Logical. Whether p-values should be added below the correlations.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A correlation table formatted for publication.
}\if{html}{\out{</div>}}
}
\description{
Generate and format a correlation table for publication.
}
\author{
Nicolas Mangin
}
