% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_univariate.R
\name{table_univariate}
\alias{table_univariate}
\title{Report univariate statistics}
\usage{
table_univariate(
  x,
  variables = NULL,
  include = c("Variable", "Count", "Min", "Median", "Mean", "Max", "Range", "St.Dev",
    "Skew", "Kurt"),
  digits = 3
)
}
\arguments{
\item{x}{Tibble or dataframe. Table with only numeric variables.}

\item{variables}{Character vector. Names of the variables.}

\item{include}{Character vector. Names of the statistics to report: "Count", "Min", "Median", "Mean", "Max", "Range", "St.Dev","Skew" and "Kurt".}

\item{digits}{Numeric. Number of digits to be displayed in the formated table.}
}
\value{
Dataframe. Table of univariate statistics.
}
\description{
Generate a table displaying univariate statistics for all the variables included in x which must be numeric.
}
\author{
Nicolas Mangin
}
