% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{trimcibt}
\alias{trimcibt}
\title{one-sample t-test on trimmed means}
\usage{
trimcibt(x, nv = 0, tr = 0.2, nboot = 100L, ci = 0.95, ...)
}
\arguments{
\item{x}{Numeric vector with or without NA's.}

\item{nv}{Test value, default to 0.}

\item{tr}{Trim for the mean, default to 0.2.}

\item{nboot}{Number of bootstrap samples for computing confidence interval for the effect size (default: 100L).}

\item{ci}{Confidence/Credible Interval (CI) level, default to 0.95 (95\%).}

\item{...}{Currently ignored.}
}
\description{
A wrapper around statsExpressions original function trimcibt
}
