% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample.R
\name{two_sample}
\alias{two_sample}
\title{Two Sample test}
\usage{
two_sample(
  data,
  x,
  y,
  rowid = NULL,
  type,
  paired = FALSE,
  var.equal = FALSE,
  effsize.type = "unbiased",
  alternative = "two.sided",
  conf.level = 0.95,
  character.only = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame from which \code{x} and \code{y} (and possibly \code{rowid} if provided) will be searched.}

\item{x}{Character name for the grouping factor. Must be present in data}

\item{y}{Character name for the response variable. Must be present in data.}

\item{rowid}{Character name for the subject-id column. If null, then is assumed that data is sorted for paired designs, creating one. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate when there are more than two levels in x and there are NAs present.}

\item{type}{Set \code{"auto"} (default) for checking the normality and homogeneity of variances for test selection. Other options are \code{"p"} for parametric, \code{"np"} for non-parametric and \code{"r"} for robust tests.}

\item{paired}{Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is \code{FALSE.}}

\item{var.equal}{Logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{effsize.type}{Options are \code{"unbiased"} or \code{"g"} for Hedges g and \code{"biased"} or \code{"d"} for Cohen's d as a measure of effect size. rank-biserial correlation is used for non-parametric analysis.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to 0.95 (95\%).}

\item{character.only}{whether to treat \code{x} as a character. Default is FALSE.}

\item{...}{Currently ignored.}
}
\description{
A list containing results from a two-sample test and effect size plus confidence intervals.
}
