% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writR-package.R
\docType{package}
\name{writR-package}
\alias{writR}
\alias{writR-package}
\title{writR: basic inferential statistics for scientific writing}
\description{
Perform and report basic inferential tests for scientific and/or technical writing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matcasti/writR}
  \item Report bugs at \url{https://github.com/matcasti/writR/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

}
\keyword{internal}
