% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_html.R
\name{add_html}
\alias{add_html}
\title{Gadget to insert various html objects in a document.}
\usage{
add_html()
}
\value{
Write lines of html codes for the desired objects where the function was called.
}
\description{
Gadget to insert various html objects in a document.
}
