% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_r.R
\name{aov_r}
\alias{aov_r}
\title{ANOVA for factorial designs}
\usage{
aov_r(
  data,
  response = NULL,
  between = NULL,
  within = NULL,
  rowid = NULL,
  ss_type = 3,
  effsize.type = "unbiased",
  sphericity = "auto",
  conf.level = 0.95,
  lbl = if (is.null(markdown)) FALSE else TRUE,
  markdown = NULL
)
}
\arguments{
\item{data}{Your dataset in long format, can have some missing values.}

\item{response}{Response variable, numeric.}

\item{between}{Character indicating the between-subject(s) factor(s)/column(s) in data. Default is NULL indicating no between-subjects factors. Must be character vector if more than one between-subject(s) factor(s)/column(s) is specified.}

\item{within}{Character indicating the within-subject(s) factor(s)/column(s) in data. Default is NULL indicating no between-subjects factors. Must be character vector if more than one within-subject(s) factor(s)/column(s) is specified.}

\item{rowid}{Character (of length 1) indicating the subject identifier column in data.}

\item{ss_type}{Type of sum of squares for repeated measures ANOVA (defaults to 3). Possible values are "II", "III", 2, or 3.}

\item{effsize.type}{The effect size used to estimate the effects of the factors on the response variable. Possible values are "eta" ("biased") or "omega" ("unbiased", the default).}

\item{sphericity}{If "none", then sphericity assumption is assumed to be met for within-subject(s) factor(s). "GG": applies Greenhouse-Geisser correction. "HF": applies Hyunh-Feldt correction. 'auto' (Default) choose the appropiate correction based on Mauchly test of sphericity (p-value > 0.05)}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to 0.95 (95\%).}

\item{lbl}{Logical (default FALSE) indicating if a report ready output is desired. This will change the output to a list with characters rather than numeric vectors.}

\item{markdown}{Logical (default FALSE). If \code{lbl} is TRUE, then this argument specify if the report-ready labels should be formated for inline code for R markdown (using mathjax and markdown syntax), or if the output should be in plain text (the default).}
}
\value{
A list with statistical results.
}
\description{
A wrapper around \code{afex::aov_ez} wich allows mix effects model.
}
\keyword{aov_r}
