% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cent_disp.R
\name{cent_disp}
\alias{cent_disp}
\title{Inline summary descriptive}
\usage{
cent_disp(
  x,
  type = if (is.null(str.a)) "auto" else "custom",
  str.a = NULL,
  k = 1,
  markdown = TRUE
)
}
\arguments{
\item{x}{Numeric vector.}

\item{type}{To choose a parametric (\code{type = 'p'}) estimate (i.e. mean and sd) or a non-parametric one (\code{type = 'np'}, median and iqr). Can be choosen automatically with \verb{type = 'auto}. It can be set to \code{'custom'} so you can insert your own expresions using glue syntax in \code{str.a} argument.}

\item{str.a}{Works only when \code{type = 'custom'}. Allows you to set your own expresions using glue syntax, e.g. \verb{str.a = "\{median\} (*MAD* = \{mad\}, *IQR* = \{IQR\})}}

\item{k}{Number of decimals.}

\item{markdown}{Whether you want the output formated for inline R markdown (TRUE) or as plain text (FALSE).}
}
\value{
A character vector of length one.
}
\description{
This function returns a character vector with mean and sd or median and iqr as a inline summary statistics for reports or articles in R markdown or plain text.
}
\keyword{cent_disp}
