% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_cell.R
\name{html_cell}
\alias{html_cell}
\title{Write html code for a cell (is ment to be used through html_layout)}
\usage{
html_cell(width = 1)
}
\arguments{
\item{width}{Integer. relative width of the cell.}
}
\value{
Character. html cell.
}
\description{
Write html code for a cell (is ment to be used through html_layout)
}
\seealso{
html_layout
}
