% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_cell.R
\name{html_cell}
\alias{html_cell}
\title{Write HTML code for a cell; it is meant to be used through html_layout.}
\usage{
html_cell(width = 1)
}
\arguments{
\item{width}{Integer. relative width of the cell.}
}
\value{
Character. HTML code for a cell.
}
\description{
Write HTML code for a cell; it is meant to be used through html_layout.
}
\seealso{
html_layout
}
