% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_star_rating.R
\name{html_star_rating}
\alias{html_star_rating}
\title{Insert a Star Rating}
\usage{
html_star_rating(source = "", page = 1)
}
\arguments{
\item{source}{Character. Name of the book.}

\item{page}{Integer. Number of the section in the book.}
}
\value{
HTML code to collect star ratings in online textbooks.
}
\description{
Function inserting a star rating in an HTML document. This requires the css and the starrating.js to be specified separately in the header and footer of the HTML document.
}
\author{
Nicolas Mangin
}
