% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_var.equal}
\alias{is_var.equal}
\title{Test for Homogeneity of Variances}
\usage{
is_var.equal(y, x, alpha = 0.05, test = "levene", center = stats::median)
}
\arguments{
\item{y}{Numeric vector (response).}

\item{x}{Grouping factor.}

\item{alpha}{Threshold for null hipotesis (of normality) rejection. Defaults to 0.05}

\item{test}{A character indicating whether to use Levene's test ("levene", the default) or Fligner's test ("fligner"). The first letter will be used.}

\item{center}{A function to compute the center of each group (valid only for Levene's Test). The mean gives the original Levene's test; the default, median, provides a more robust test.}
}
\value{
A logical of length one, indicating if the p.value of the test was greater than the alpha (which defaults to 0.05).
}
\description{
Test for Homogeneity of Variances
}
