% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample.R
\name{one_sample}
\alias{one_sample}
\title{One Sample test}
\usage{
one_sample(
  data,
  y,
  type,
  test.value = 0,
  effsize.type = "unbiased",
  alternative = "two.sided",
  conf.level = 0.95,
  character.only = FALSE,
  markdown,
  ...
)
}
\arguments{
\item{data}{data frame from which \code{x} and \code{y} (and possibly \code{rowid} if provided) will be searched.}

\item{y}{Character for the response variable Must be present in data}

\item{type}{set \code{"auto"} (default) for checking the normality for test selection. Other options are \code{"p"} for parametric, \code{"np"} for non-parametric and \code{"r"} for robust tests.}

\item{test.value}{A number indicating the true value of the mean (Default: 0) to be tested.}

\item{effsize.type}{options are \code{"unbiased"} or \code{"g"} for Hedges g and \code{"biased"} or \code{"d"} for Cohen's d as a measure of effect size. rank-biserial correlation is used for non-parametric analysis.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to 0.95 (95\%).}

\item{character.only}{whether to treat \code{x} as a character. Default is FALSE.}

\item{markdown}{Logical (default FALSE). If \code{lbl} is TRUE, then this argument specify if the report-ready labels should be formated for inline code for R markdown (using mathjax and markdown syntax), or if the output should be in plain text (the default).}

\item{...}{Currently ignored.}
}
\description{
A list containing results from a one-sample test.
}
