% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_two_sample.R
\name{pairs_two_sample}
\alias{pairs_two_sample}
\title{Pairwise Two Sample test}
\usage{
pairs_two_sample(
  data,
  x,
  y,
  rowid = NULL,
  type,
  paired = FALSE,
  var.equal = FALSE,
  effsize.type = "unbiased",
  alternative = "two.sided",
  conf.level = 0.95,
  markdown,
  character.only = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame from which \code{x} and \code{y} (and possibly \code{rowid} if provided) will be searched.}

\item{x}{Character for the grouping factor. Must be present in data}

\item{y}{Character for the response variable. Must be present in data.}

\item{rowid}{Character for the subject-id column. If null, then is assumed that data is sorted for paired designs, creating one. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate when there are more than two levels in x and there are NAs present.}

\item{type}{Set \code{"auto"} (default) for checking the normality and homogeneity of variances for test selection. Other options are \code{"p"} for parametric, \code{"np"} for non-parametric and \code{"r"} for robust tests.}

\item{paired}{Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is \code{FALSE.}}

\item{var.equal}{Logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{effsize.type}{Options are \code{"unbiased"} or \code{"g"} for Hedges g and \code{"biased"} or \code{"d"} for Cohen's d as a measure of effect size for parametric test. The rank-biserial correlation is used for non-parametric analysis.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to 0.95 (95\%).}

\item{markdown}{Logical (default FALSE). If \code{lbl} is TRUE, then this argument specify if the report-ready labels should be formated for inline code for R markdown (using mathjax and markdown syntax), or if the output should be in plain text (the default).}

\item{character.only}{Logical. checks whether to use the unevaluated expression or its
content (when TRUE), asumming is a character vector. Defaults to \code{FALSE}.}

\item{...}{Currently ignored.}
}
\description{
A wrapper around \code{two_sample()} for pairwise comparisons. For posthoc testing we recomend using \code{pairwise_test()} accordingly.
}
