% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.R
\name{rcl}
\alias{rcl}
\title{Reshape data from long format}
\usage{
rcl(data, variable, by, paired = FALSE, spread = FALSE)
}
\arguments{
\item{data}{Your dataset in long format, can have missing values.}

\item{variable}{Response variable, numeric.}

\item{by}{Grouping variable, a factor. It can have more than two levels.}

\item{paired}{If the design is for paired groups, eliminates missing values at subject level (default FALSE).}

\item{spread}{TRUE will give you wide format output, otherwise long format (default FALSE).}
}
\value{
A dataframe with NAs removed.
}
\description{
This function allows you to handle missing values and is a wrapper around \code{ipmisc::long_to_wide_converter()}.
}
\keyword{rcl}
