% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_report.R
\name{stat_report}
\alias{stat_report}
\title{Add stars to a vector of estimateicients based on the p-value.}
\usage{
stat_report(
  coefficient = "$\\beta$",
  estimate = NULL,
  std.error = NULL,
  threshold = 0.05,
  digits = 3,
  stats = c("cf", "ci", "pv")
)
}
\arguments{
\item{coefficient}{Character. Name or identifier of the coefficient.}

\item{estimate}{Numeric. Estimate of the coefficient.}

\item{std.error}{Numeric. Standard error arounf the estimate.}

\item{threshold}{Numeric. Probability of 0 contained in the confidence interval.}

\item{digits}{Integer. Number of digits for rounding.}

\item{stats}{Character vector. Statistics to report: "cf" for the coefficient, "ci" for the confidence internal, "pv" for the p-value}
}
\value{
Character vector with coefficients formatted for publication.
}
\description{
Add stars to a vector of estimateicients based on the p-value.
}
