% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample.R
\name{two_sample}
\alias{two_sample}
\title{Two Sample test}
\usage{
two_sample(
  data,
  x,
  y,
  rowid = NULL,
  type = "auto",
  paired = FALSE,
  var.equal = FALSE,
  trim = 0.2,
  nboot = 100L,
  effsize.type = "unbiased",
  alternative = "two.sided",
  conf.level = 0.95,
  lbl = if (is.null(markdown)) FALSE else TRUE,
  markdown = NULL,
  internal = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame from which \code{x} and \code{y} (and possibly \code{rowid} if provided) will be searched.}

\item{x}{Character name for the grouping factor. Must be present in data}

\item{y}{Character name for the response variable. Must be present in data.}

\item{rowid}{Character name for the subject-id column. If null, then is assumed that data is sorted for paired designs, creating one. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate when there are more than two levels in x and there are NAs present.}

\item{type}{Set \code{"auto"} (default) for checking the normality and homogeneity of variances for test selection. Other options are \code{"p"} for parametric, \code{"np"} for non-parametric and \code{"r"} for robust tests.}

\item{paired}{Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is \code{FALSE.}}

\item{var.equal}{Logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{trim}{Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.}

\item{nboot}{Number of bootstrap samples for computing confidence interval for the effect size (Default: 100L).}

\item{effsize.type}{Options are \code{"unbiased"} or \code{"g"} for Hedges g and \code{"biased"} or \code{"d"} for Cohen's d as a measure of effect size. rank-biserial correlation is used for non-parametric analysis.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to 0.95 (95\%).}

\item{lbl}{Logical (default FALSE) indicating if a report ready output is desired. This will change the output to a list with characters rather than numeric vectors.}

\item{markdown}{Logical (default FALSE). If \code{lbl} is TRUE, then this argument specify if the report-ready labels should be formated for inline code for R markdown (using mathjax and markdown syntax), or if the output should be in plain text (the default).}

\item{internal}{Logical to whether this function is being used inside of internal functions.}

\item{...}{Currently ignored.}
}
\description{
A list containing results from a two-sample test and effect size plus confidence intervals.
}
