% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writR-package.R
\docType{package}
\name{writR-package}
\alias{writR}
\alias{writR-package}
\title{writR: Inferential statistics and reporting in APA style}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Perform and report basic inferential tests for group differences in APA style for articles writings in Markdown or plain text format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matcasti/writR}
  \item Report bugs at \url{https://github.com/matcasti/writR/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

}
\keyword{internal}
